// pos-server/src/pos/dto/t-products-purchase.dto.ts
export class TProductsPurchaseDto {

    invoiceNo: string;

    productId: number;

    productCode: string;

    productName: string;
        
    unitId: number;

    unitName: string;

    childUnitId: number;

    childUnitName: string;

    unitCost: number;

    qty: number;

    promoQty: number;

    promoStatus: string;

    cost: number;

    sellPrice: number;
   
    expDate: Date;

}